Extract gameplay branch commitments
