+++ crates/oasis7/src/viewer/runtime_live/gameplay_snapshot.rs
@@ -3,13 +3,13 @@ use crate::runtime::{
     WorldEventBody as RuntimeWorldEventBody, WorldState,
 };
 use crate::simulator::persist::{
-    PlayerAgentClaimSnapshot, PlayerGameplayAction, PlayerGameplayBranchCommitment,
-    PlayerGameplayCausalityKind, PlayerGameplayExecutionState, PlayerGameplayGoalKind,
-    PlayerGameplayRecentFeedback, PlayerGameplaySnapshot, PlayerGameplayStageId,
-    PlayerGameplayStageStatus,
+    PlayerAgentClaimSnapshot, PlayerGameplayAction, PlayerGameplayCausalityKind,
+    PlayerGameplayExecutionState, PlayerGameplayGoalKind, PlayerGameplayRecentFeedback,
+    PlayerGameplaySnapshot, PlayerGameplayStageId, PlayerGameplayStageStatus,
 };
 use crate::viewer::{ControlCompletionAck, ControlCompletionStatus, ViewerControl};
 
+use super::branch_commitment::{branch_recommendations, effective_branch_stage_status};
 pub(super) use super::gameplay_snapshot_helpers::apply_runtime_snapshot_empty_entities_blocker;
 use super::gameplay_snapshot_helpers::{
     base_available_actions, blocker_next_step, primary_factory_for_player_gameplay,
@@ -485,104 +485,6 @@ fn finalize_player_gameplay_snapshot(
     gameplay.resume_next_step = Some(gameplay.next_step_hint.clone());
     apply_small_player_lane_truth(&mut gameplay);
     gameplay
-}
-
-fn branch_recommendations(
-    industry_stage: IndustryStage,
-    stage_status: PlayerGameplayStageStatus,
-    available_actions: &[PlayerGameplayAction],
-) -> Vec<PlayerGameplayBranchCommitment> {
-    if stage_status != PlayerGameplayStageStatus::BranchReady {
-        return Vec::new();
-    }
-
-    let enabled = |action_id: &str| {
-        available_actions
-            .iter()
-            .any(|action| action.action_id == action_id && action.disabled_reason.is_none())
-    };
-    let commitment =
-        |action_id: &str,
-         route_label: &str,
-         immediate_gain: &str,
-         future_beat_changed: &str,
-         risk_or_lockin: &str,
-         next_session_hook: &str| PlayerGameplayBranchCommitment {
-            action_id: action_id.to_string(),
-            route_label: route_label.to_string(),
-            immediate_gain: immediate_gain.to_string(),
-            future_beat_changed: future_beat_changed.to_string(),
-            risk_or_lockin: risk_or_lockin.to_string(),
-            next_session_hook: next_session_hook.to_string(),
-        };
-
-    let scale_out_candidates = [
-        commitment(
-            "schedule_recipe_smelter_alloy_plate",
-            "Deepen the smelter line",
-            "Convert the stable line into higher-tier alloy output.",
-            "The next beat shifts from basic throughput to advanced inputs.",
-            "Consumes line time that could have supplied basic materials.",
-            "Return to turn the alloy output into the next production upgrade.",
-        ),
-        commitment(
-            "build_factory_assembler_mk1",
-            "Open a second production line",
-            "Add the assembler capability and widen the real recipe surface.",
-            "The next beat becomes coordinating two complementary factories.",
-            "Commits construction materials before the new line produces value.",
-            "Return to choose and run the assembler's first recipe.",
-        ),
-    ];
-    let governance_candidates = [
-        commitment(
-            "schedule_recipe_assembler_module_rack",
-            "Produce a module rack",
-            "Create a governance-stage finished component.",
-            "The next beat shifts toward assembling durable industrial modules.",
-            "Uses advanced assembler inputs and occupies the current recipe slot.",
-            "Return to connect the module rack to a factory-core production run.",
-        ),
-        commitment(
-            "schedule_recipe_assembler_factory_core",
-            "Produce a factory core",
-            "Create infrastructure output from the mature assembler line.",
-            "The next beat shifts toward deploying another industrial capability.",
-            "Consumes a module rack and alloy plate before yielding the core.",
-            "Return to decide where the new factory core creates the most leverage.",
-        ),
-        commitment(
-            "schedule_recipe_smelter_alloy_plate",
-            "Sustain alloy throughput",
-            "Produce another alloy plate through the mature smelter line.",
-            "The next beat keeps advanced assembler inputs supplied while other routes unlock.",
-            "Uses the smelter queue and iron inputs without automatically advancing governance.",
-            "Return to inspect the alloy output and choose the next advanced recipe.",
-        ),
-    ];
-
-    let candidates = match industry_stage {
-        IndustryStage::ScaleOut => scale_out_candidates.as_slice(),
-        IndustryStage::Governance => governance_candidates.as_slice(),
-        IndustryStage::Bootstrap => return Vec::new(),
-    };
-    candidates
-        .into_iter()
-        .filter(|candidate| enabled(candidate.action_id.as_str()))
-        .take(3)
-        .cloned()
-        .collect()
-}
-
-fn effective_branch_stage_status(
-    stage_status: PlayerGameplayStageStatus,
-    recommendations: &[PlayerGameplayBranchCommitment],
-) -> PlayerGameplayStageStatus {
-    if stage_status == PlayerGameplayStageStatus::BranchReady && recommendations.is_empty() {
-        PlayerGameplayStageStatus::Blocked
-    } else {
-        stage_status
-    }
 }
 
 pub(super) fn build_player_gameplay_snapshot(
@@ -1293,34 +1195,4 @@ pub(super) fn build_player_gameplay_snapshot(
         rebuild_available: None,
         pivot_available: None,
     })
-}
-
-#[cfg(test)]
-mod branch_commitment_tests {
-    use super::*;
-
-    #[test]
-    fn branch_ready_downgrades_when_every_published_candidate_is_disabled() {
-        let actions = vec![PlayerGameplayAction {
-            action_id: "schedule_recipe_smelter_alloy_plate".to_string(),
-            label: "Schedule alloy plate".to_string(),
-            protocol_action: "gameplay_action.submit".to_string(),
-            target_agent_id: None,
-            disabled_reason: Some("missing iron ingot".to_string()),
-        }];
-        let recommendations = branch_recommendations(
-            IndustryStage::ScaleOut,
-            PlayerGameplayStageStatus::BranchReady,
-            actions.as_slice(),
-        );
-
-        assert!(recommendations.is_empty());
-        assert_eq!(
-            effective_branch_stage_status(
-                PlayerGameplayStageStatus::BranchReady,
-                recommendations.as_slice(),
-            ),
-            PlayerGameplayStageStatus::Blocked,
-        );
-    }
 }
